requirements.txt
anyio==4.9.0
certifi==2025.1.31
easyocr==1.7.2
filelock==3.18.0
flet==0.27.6
flet-desktop==0.27.6
fsspec==2025.3.2
h11==0.14.0
httpcore==1.0.8
httpx==0.28.1
idna==3.10
imageio==2.37.0
Jinja2==3.1.6
lazy_loader==0.4
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
ninja==1.11.1.4
numpy==2.2.4
oauthlib==3.2.2
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
packaging==24.2
pillow==11.2.1
pyclipper==1.3.0.post6
pytesseract==0.3.13
python-bidi==0.6.6
python-dateutil==2.9.0.post0
PyYAML==6.0.2
repath==0.9.0
scikit-image==0.25.2
scipy==1.15.2
setuptools==78.1.0
shapely==2.1.0
six==1.17.0
sniffio==1.3.1
sympy==1.13.1
tifffile==2025.3.30
torch==2.6.0
torchvision==0.21.0
typing_extensions==4.13.2
